(chores) came-jms: avoid failing the test due to session being closed
This reuses the same approach from another commit and applies to another similar test (ref.: 2396b3c4b8d7f1df0b411330e5497fb16bc669d0)
