+++ components/camel-jms/src/test/java/org/apache/camel/component/jms/JmsInOnlyDisableTimeToLiveTest.java
@@ -22,14 +22,13 @@ import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.test.infra.core.CamelContextExtension;
-import org.apache.camel.test.infra.core.DefaultCamelContextExtension;
+import org.apache.camel.test.infra.core.TransientCamelContextExtension;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.jms.IllegalStateException;
 
 public class JmsInOnlyDisableTimeToLiveTest extends AbstractJMSTest {
 
@@ -37,7 +36,7 @@ public class JmsInOnlyDisableTimeToLiveTest extends AbstractJMSTest {
 
     @Order(2)
     @RegisterExtension
-    public static CamelContextExtension camelContextExtension = new DefaultCamelContextExtension();
+    public static CamelContextExtension camelContextExtension = new TransientCamelContextExtension();
     protected CamelContext context;
     protected ProducerTemplate template;
     protected ConsumerTemplate consumer;
@@ -47,9 +46,7 @@ public class JmsInOnlyDisableTimeToLiveTest extends AbstractJMSTest {
 
     @Test
     public void testInOnlyExpired() throws Exception {
-        MyCoolBean cool = new MyCoolBean();
-        cool.setProducer(template);
-        cool.setConsumer(consumer);
+        MyCoolBean cool = new MyCoolBean(consumer, template, "JmsInOnlyDisableTimeToLiveTest");
 
         getMockEndpoint("mock:result").expectedBodiesReceived("World 1");
 
@@ -72,9 +69,7 @@ public class JmsInOnlyDisableTimeToLiveTest extends AbstractJMSTest {
 
     @Test
     public void testInOnlyDisabledTimeToLive() throws Exception {
-        MyCoolBean cool = new MyCoolBean();
-        cool.setProducer(template);
-        cool.setConsumer(consumer);
+        MyCoolBean cool = new MyCoolBean(consumer, template, "JmsInOnlyDisableTimeToLiveTest");
 
         getMockEndpoint("mock:result").expectedBodiesReceived("World 2");
 
@@ -130,46 +125,6 @@ public class JmsInOnlyDisableTimeToLiveTest extends AbstractJMSTest {
         context = camelContextExtension.getContext();
         template = camelContextExtension.getProducerTemplate();
         consumer = camelContextExtension.getConsumerTemplate();
-    }
-
-    public static class MyCoolBean {
-        private int count;
-        private ConsumerTemplate consumer;
-        private ProducerTemplate producer;
-
-        public void setConsumer(ConsumerTemplate consumer) {
-            this.consumer = consumer;
-        }
-
-        public void setProducer(ProducerTemplate producer) {
-            this.producer = producer;
-        }
-
-        public void someBusinessLogic() {
-            // loop to empty queue
-            while (true) {
-                // receive the message from the queue, wait at most 2 sec
-                try {
-                    String msg = consumer.receiveBody("activemq:JmsInOnlyDisableTimeToLiveTest.in", 2000, String.class);
-                    if (msg == null) {
-                        // no more messages in queue
-                        break;
-                    }
-                    // do something with body
-                    msg = "Hello " + msg;
-
-                    // send it to the next queue
-                    producer.sendBodyAndHeader("activemq:JmsInOnlyDisableTimeToLiveTest.out", msg, "number", count++);
-                } catch (IllegalStateException e) {
-                    if (e.getCause() instanceof jakarta.jms.IllegalStateException) {
-                        // session is closed
-                        LOG.warn("JMS Session is closed");
-                        break;
-                    }
-                }
-
-            }
-        }
     }
 
 }
